Add script to download Valentine's data
Wrote a utility script to download Valentine Svensson's database.
It also takes a 10-dataset random subset meeting certain criteria. It
does this because I want a small subset of suitable datasets to develop
the first structure and utility scripts for the single-cell-database.
The script also datetime-stamps its downloaded files so that there is
no overwriting of previous data downloads.