Add import method for devel_data and refactor
I refactored the devel_data download script into a utility
library for working with the devel_data. I added an import
function, so that the parsing doesn't have to happen every
time I want to do something with it. I also moved it from the
utils folder to src so I don't have to deal with imports across
folders.