Add get_timestamp() utility method
Added a new file for general helper methods. First method added
was a method to generate datestamps, timestamps, or a combination.
They are generated in strftime() format '%y%m%d', '%H%M%S', or both
concatenated with a '_'. Also refactored download_devel_data() in
devel_data.py to use the new method instead of hardcoding it.