Add methods to create a new database entry folder
There is some small documentation refactoring. Otherwise, the
main changes involve adding some methods for scraping the
SOFT file and the supplementary files for a given GEO Series,
assigning the new database entry a UUID, and placing everything
in a new folder, along with a log file. There are helper
methods included that enable writing certain things to the
log file.