Refactor a lot & add a new_entry interface
I completely refactored this library. There are now 4 global
constants that get used for most of the UI aspects:
    _PATH_TO_DATA: Hardcoded path to the external_metadata.tsv
                   file
    _COLUMN_DESCRIPTIONS: Dictionary of helpful descriptions
                          for each column, suitable for printing
    _COLUMN_INDEX: Dictionary of 0-based indices, indicating the
                   column order in the file
    _COLUMN_MANDATORY: Dictionary of booleans, indicating
                       whether or not the column is mandatory
Additionally, I added a large new function, get_new_row_input()
that presents a command line interface for a user to input
a new entry. It returns a list of values, it doesn't actually
enter the new entry.