Enable adding a CellRanger entry to the database
Add functions to be used during a manual entry of a new Cell
Ranger dataset in the database. These functions should be
run in a single interactive Python session, with minor prep
being done in between each step in a separate shell. For
example, the relevant files may be named appropriately
after downloading, but before loom file creation.