Move all file connections to with-blocks
Edited all code that used file connections, either locally
or from a getter in access.py to manage the file connection
with a "with" block. This ensures that file connections get
closed if a method throws an error. This also holds true
even for connections passed from a getter, even if they are
renamed within the with statement. I also fixed a little
bug in some setters for internal metadata where a column
might not be added to the columns_written dict in
internal_metadata.py

with-block example:

>>> def set_metadata(lfile, return_early = False,
...                  error_out = False):
...    with lfile as f:
...        f.write('Line 1 of metadata.')
...        if error_out:
...            raise AssertionError('Here is an error!')
...        if return_early:
...            return('I returned early!')
...        f.write('Line 2 of metadata.')
...    return('I returned normally!')
...
>>> lfile = h5py.File('expr_mat.loom', 'r+')
>>> set_metadata(lfile)
'I returned normally!'
>>> lfile.closed
True
>>> lfile = h5py.File('expr_mat.loom', 'r+')
>>> set_metadata(lfile, return_early = True)
'I returned early!'
>>> lfile.closed
True
>>> lfile = h5py.File('expr_mat.loom', 'r+')
>>> set_metadata(lfile, error_out = True)
AssertionError: Here is an error!
>>> lfile.closed
True