Added dict_to_dataframe and a yes/no getter
Added 2 utilities. One converts a dict of dicts to a pandas
dataframe. I'm using this as the second half of a SOFT file
parser for SAMPLE characteristics. The second utility is a
getter for a yes/no response from a user that is robust to
invalid responses. Also updated pretty_str_list(). It doesn't
print anymore; it only returns the string. It also has a new
option to force one-element-per-line.