Update the schema: no mappings and new names
This is an update of the global constants to match the new
schema. There are no more mappings. Additionally, the constants
are named CELL and GENE instead of COL and ROW. This is because
the author_annot schema uses those labels. The author_annot uses
those labels because AnnData puts cells on rows and vice versa,
thus the col_author_annot and row_author_annot variable names
in the adata['uns'].keys() were incorrect. Better to be specific
about the variable, not the place it may or may not be in.