Add a SOFT file parser and update a few details
I added a SOFT file parser that scrapes SAMPLE characteristics
into a dict of dicts. Meant to be used with the dict of dicts
converter in general_utils.py. Also added a couple of
convenience functions in add_dataset.py so I can quickly
reference the internal universal metadata schema and the
tissue list. Finally, updated a few minor details and changed
some methods to use the new yes/no getter in general utils
instead of implementing in each place.