Completely overhauled access.R + minor updates
access.py and external_metadata got a couple of minor updates,
and access.py got a new method called get_expr_mat_names(),
which just allows access to the list of available matrices.
access.R got updated to be completely in sync with access.py.
This was a lot of new methods, documentation, and rewriting of
the get_sce() method.