Update Accession/Gene handling
Rewrites a number of functions to handle the Accession and Gene
row_attrs in a smarter way. The problem is that only one has
to exist, and they aren't included in the global constants.
All functions now prefer Accession over Gene, will refuse to
write Accession if it has non-unique values, and will not use
either as an Index on read if they have non-unique values. All
access functions will check to see if either are being included
as metadata and not as an index, and if so will insert them at
the front of the sort order.