Make this a package for local users
This commit makes this codebase available to anyone else on
the Gillis Lab servers as an importable Python package. It
adds a real __init__.py file, fixes internal imports to act
with relative imports, renames an old file, and adds a setup.sh
file. The setup.sh file naively adds the appropriate folder
to the user's PYTHONPATH environment variable. This allows
multiple versions of this codebase.