Refactor to better handle varying pathnames
The tyrone mount is not called /data on all servers. On dactyl,
it's /tyronedata. This makes storing filepaths in the external
metadata very inconvenient. This commit refactors the whole
codebase to dynamically generated filepaths from UUIDs and the
global path to the database. Then, it inserts a hostname check
for dactyl. This introduces a single choke point to adjust
filepath generation for all access functions.