Fix a collection of bugs found during testing
This commit mostly fixes a myriad of bugs discovered during
testing. Three functionality changes occurred as a result:
1) author_annot HDF5 Groups now Must have a column_order
attribute, even if there are no columns. 2) getters for
all internal metadata now return empty dataframes instead
of None/NULL when there are no columns. This is to better
signal the status of the dataset, as well as to show the IDs.
3) the create_loom.get_expr_matrix_from_cellranger() function
now can return a dataframe of features or barcodes if the tsv
file contains extra info (e.g. accession IDs and gene names).