Add 2 new utility functions and fix bugs
I added 2 new functions to general_utils.py. The first one is
similar to get_yes_or_no(). It provides a robust way to get
a text command from the user on the command line. The second one
is a special case of pretty_str_list. It uses pretty_str_list()
to create a left-aligned width-constrained pretty printable
version of a long string. Creating this resulted in a redesign
of pretty_str_list(), but the input/output is the same.