Implement a major overhaul of multiple schema.
I made several changes to the schema. This commit implements
all of it. The first change (1) removes all the ambiguity
around "Gene" and "Accession" in the internal metadata. Both
are required fields and should be filled with "-1" if they are
missing. This simplified a lot of code that danced around the
fact that neither field existed in the gene universal internal
metadata schema. The CellID field is still special, but doesn't
require as many dances. The second change (2) adds an HDF5
attribute to ALL internal metadata columns. It is called
"all_missing" and it indicates whether or not the entire column
is "-1". This enables much quicker checks because the entire
column doesn't have to be traversed. The third change (3) adds
a column to the external metadata called "author_clusters". This
indicates whether or not the author(s) provided cell clusters,
annotated or otherwise. Finally, there are 2 smaller changes.
The first one updated get_anndata() in access.py to allow for
dropping missing internal universal metadata. The second added
a function in create_loom.py to get mat, bar, feat from a
different Cell Ranger output: filtered_feature_bc_matrix_h5.h5