Fix a bug with rownames in get_sce()/update docs
Rownames were not being set correctly in get_sce() because of
loom file closures in helper methods. Additionally, the rowData
construction was changed, so the rownames construction process
had to change to reflect this update. Additionally, I added
a small section in the docs to set a standard for choosing
dataset delineations.